fix for ERB's newline not working on heroku
For some reason, a single line substition inside an each block in ERB
isn't coming with proper line breaks on heroku. This issue is not
reproducible when trying to run with production on local machine.
Perhaps an issue with different versions of Ruby/ERB. This commit is
just a quick fix.